Move database actions into shared
