+++ enterprise/frontend/src/metabase-enterprise/google_drive/GdriveConnectionModal.tsx
@@ -2,12 +2,12 @@ import dayjs from "dayjs";
 import { type FormEvent, useState } from "react";
 import { c, t } from "ttag";
 
-import { reloadSettings } from "metabase/admin/settings/settings";
 import { skipToken, useGetUserQuery } from "metabase/api";
 import { CopyButton } from "metabase/common/components/CopyButton";
 import { ExternalLink } from "metabase/common/components/ExternalLink";
 import { Markdown } from "metabase/common/components/Markdown";
 import { useDispatch } from "metabase/redux";
+import { reloadSettings } from "metabase/redux/settings";
 import {
   Box,
   Button,
