+++ frontend/src/metabase/home/components/CustomHomePageModal/CustomHomePageModal.tsx
@@ -2,11 +2,11 @@ import { useCallback, useState } from "react";
 import { jt, t } from "ttag";
 
 import { trackCustomHomepageDashboardEnabled } from "metabase/admin/settings/analytics";
-import { updateSettings } from "metabase/admin/settings/settings";
 import { DashboardSelector } from "metabase/common/components/DashboardSelector/DashboardSelector";
 import { Link } from "metabase/common/components/Link";
 import CS from "metabase/css/core/index.css";
 import { useDispatch } from "metabase/redux";
+import { updateSettings } from "metabase/redux/settings";
 import { addUndo, dismissUndo } from "metabase/redux/undo";
 import { refreshCurrentUser } from "metabase/redux/user";
 import { Box, Button, Flex, Modal, Text } from "metabase/ui";
