+++ frontend/src/metabase/redux/settings.ts
@@ -1,6 +1,12 @@
-import { createAction, createReducer } from "@reduxjs/toolkit";
+import {
+  type ThunkDispatch,
+  type UnknownAction,
+  createAction,
+  createReducer,
+} from "@reduxjs/toolkit";
 
 import { sessionApi } from "metabase/api";
+import type { State } from "metabase/redux/store";
 import { createAsyncThunk, createThunkAction } from "metabase/redux/utils";
 import { SettingsApi } from "metabase/services";
 import type { Settings, UserSettings } from "metabase-types/api";
@@ -73,6 +79,42 @@ export const updateSetting = createThunkAction(
         throw error;
       } finally {
         await dispatch(refreshSiteSettings());
+      }
+    };
+  },
+);
+
+export const reloadSettings =
+  () => async (dispatch: ThunkDispatch<State, unknown, UnknownAction>) => {
+    await dispatch(refreshSiteSettings());
+  };
+
+export const INITIALIZE_SETTINGS =
+  "metabase/admin/settings/INITIALIZE_SETTINGS";
+export const initializeSettings = createThunkAction(
+  INITIALIZE_SETTINGS,
+  () => async (dispatch) => {
+    try {
+      await dispatch(reloadSettings());
+    } catch (error) {
+      console.error("error fetching settings", error);
+      throw error;
+    }
+  },
+);
+
+export const UPDATE_SETTINGS = "metabase/admin/settings/UPDATE_SETTINGS";
+export const updateSettings = createThunkAction(
+  UPDATE_SETTINGS,
+  function (settings) {
+    return async function (dispatch) {
+      try {
+        await SettingsApi.putAll(settings);
+      } catch (error) {
+        console.error("error updating settings", settings, error);
+        throw error;
+      } finally {
+        await dispatch(reloadSettings());
       }
     };
   },
